Реализовать ФЗ6 2024.07.27
