+++ src/service/HistoryManager.java
@@ -7,5 +7,7 @@ public interface HistoryManager {
 
     void add(Task task);
 
+    void remove(int id);
+
     List<Task> getHistory();
 }
