+++ src/service/TaskManager.java
@@ -30,4 +30,6 @@ public interface TaskManager {
 
     List<Task> getHistory();
 
+    void removeTaskFromHistory(int id);
+
 }
